use sample snippets
